Remove leading spaces
I noticed that some of the text that was supposed to be in bold in your readme was not, due to leading spaces which I assume were introduced by a hard line-wrap rule. I've moved the text around to try and make sure it all fits inside a hard wrap.